Refatorando app

+++ app/Core/Model.php
@@ -9,8 +9,10 @@ abstract class Model{
     private $where = [];
     private $__data = [];
     protected $pk = 'id';
-
     private $__storage = false;
+    protected $__protected_delete = false;
+    
+    private $__protected_delete_column = 'exclusao_data';
     public function __construct($id = null){
         if(isset($id)){
             $this->load($id);
@@ -97,11 +99,17 @@ abstract class Model{
     }
 
     public function delete(){
+        
         if($this->__storage){
-           $sql = "DELETE FROM $this->table WHERE $this->pk = :$this->pk;";
-           $this->query($sql, [$this->pk => $this->{$this->pk}]);
-           $this->__storage = false;
-           return true;
+            if($this->__protected_delete){
+                $this->update([$this->__protected_delete_column => date('Y-m-d H:i:s')]);
+            }else{
+                $sql = "DELETE FROM $this->table WHERE $this->pk = :$this->pk;";
+                $this->query($sql, [$this->pk => $this->{$this->pk}]);
+            }
+
+            $this->__storage = false;
+            return true;
         }   
         return false;
     }
@@ -143,12 +151,22 @@ abstract class Model{
         $data = [];
 
         if(count($this->where)>0){
-            $this->where[0][0] = 'WHERE';
+            $this->where[0][0] = '';
             foreach($this->where as $key => $w){
                 $where .= " $w[0] $w[1] $w[2] :w$key";
                 $data["w$key"] = $w[3]; 
             }
             $this->where = [];
+        }
+   
+        if($this->__protected_delete){
+            if(empty($where)){
+                $where = " WHERE $this->__protected_delete_column IS NULL";
+            }else{
+                $where = " WHERE ($this->__protected_delete_column IS NULL) AND ($where)";
+            }
+        }else if(!empty($where)){
+            $where = " WHERE $where";
         }
         return [$where, $data];
     }
